refactored input plugins, improved coding style
