+++ flowcache.h
@@ -112,6 +112,7 @@ public:
    {
       export_queue = queue;
    }
+
    virtual void export_expired(time_t ts)
    {
    }
