+++ nfbCInterface/include/ndpreader.h
