+++ nfbCInterface/include/ndpreader.hpp
